mastodon-fetcher/pyproject.toml
[project]
name = "news-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Mastodon.py",
    "ipykernel",
    "kafka-python"
]
